+++ apps/imapi/src/fts_logic.erl
@@ -25,8 +25,8 @@ user_search_page(_, Page, Size, <<>>) ->
     imboy_response:page_payload(0, Page, Size, []);
 user_search_page(Uid, Page, Size, Keywrod) ->
     Offset = (Page - 1) * Size,
-    Total = fts_repo:count_for_user_search_page(Keywrod),
-    case fts_repo:user_search_page(Keywrod, Size, Offset) of
+    Total = fts_user_repo:count_for_user_search_page(Keywrod),
+    case fts_user_repo:user_search_page(Keywrod, Size, Offset) of
         {ok, _, []} ->
             imboy_response:page_payload(Total, Page, Size, []);
         {ok, ColumnLi, Items0} ->